Clean up include statements and move more headers to the cpp (cherry picked from commit 9a98bd5)
