+++ src/interop/logic/plot/plot_sample_qc.cpp
@@ -7,6 +7,7 @@
  */
 #include "interop/logic/plot/plot_sample_qc.h"
 
+#include "interop/util/math.h"
 #include "interop/logic/utils/enums.h"
 
 namespace illumina { namespace interop { namespace logic { namespace plot {
