+++ src/interop/model/metrics/tile_metric.cpp
@@ -11,6 +11,7 @@
 
 #include "interop/model/metrics/tile_metric.h"
 #include "interop/io/format/metric_format_factory.h"
+#include "interop/util/math.h"
 
 using namespace illumina::interop::model::metrics;
